Latest master from upstream

+++ lib/ghtorrent/commands/ght_load.rb
@@ -41,7 +41,7 @@ Loads data from a MongoDB collection or a file to a queue for further processing
     options.opt :number, 'Total number of items to load',
                 :short => 'n', :type => :int, :default => 2**48
     options.opt :rate, 'Number of items to load per second',
-                :type => :int, :default => 1000
+                :type => :float, :default => 1000.0
     options.opt :route_key, 'Routing key to attached to loaded items', :type => String
   end
 
@@ -101,7 +101,15 @@ Loads data from a MongoDB collection or a file to a queue for further processing
   end
 
   def file_stream
-    File.open(options[:file])
+    if File.exists? @last_load_file
+      last = File.open(@last_load_file, &:readline)
+      lines = File.open(options[:file]).readlines
+      idx = lines.find_index{|i| i == last}
+      puts "Skipping #{idx + 1} lines up to #{last.strip}. Remove #{@last_load_file} to avoid."
+      lines[idx..-1]
+    else
+      File.open(options[:file])
+    end
   end
 
   def file_process(e)
@@ -114,14 +122,16 @@ Loads data from a MongoDB collection or a file to a queue for further processing
       @mode = :file
       alias :process :file_process
       alias :stream :file_stream
+      @last_load_file = "#{options[:file]}.lastload"
     else
       @mode = :mongodb
       alias :process :mongo_process
       alias :stream :mongo_stream
+      @last_load_file = "events.lastload"
     end
 
     # Num events read
-    total_read = current_sec_read = 0
+    total_read = current_min_read = 0
 
     conn = Bunny.new(:host => config(:amqp_host),
                      :port => config(:amqp_port),
@@ -152,21 +162,26 @@ Loads data from a MongoDB collection or a file to a queue for further processing
 
           # Basic rate limiting
           if options[:rate_given]
-            current_sec_read += 1
-            if current_sec_read >= options[:rate]
+            current_min_read += 1
+            if current_min_read >= options[:rate] * 60
               time_diff = (Time.now - ts) * 1000
-              if time_diff <= 1000.0
-                puts "Rate limit reached, sleeping for #{1000 - time_diff} ms"
-                sleep((1000.0 - time_diff) / 1000)
+              if time_diff <= 60 * 1000.0
+                puts "Rate limit reached, sleeping for #{60 * 1000 - time_diff} ms"
+                File.open(@last_load_file,'w'){|f| f.puts id}
+                sleep((60 * 1000.0 - time_diff) / 1000)
               end
-              current_sec_read = 0
+              current_min_read = 0
               ts = Time.now
             end
           end
 
+          if total_read % 1000 == 0
+            File.open(@last_load_file,'w'){|f| f.puts id}
+          end
+
           if total_read >= options[:number]
             puts 'Finished reading, exiting'
-            return 
+            return
           end
         end
         stopped = true
